+++ HW1/hw1.h
@@ -5,11 +5,11 @@
 #include <stdlib.h>
 #include <math.h>
 
-float* filter_signal(float* signal, float* filter, int signal_size, int filter_size);
-void modulate(float* signal, int signal_len, float fs, float omega);
+double* filter_signal(double* signal, double* filter, int signal_size, int filter_size);
+void modulate(double* signal, int signal_len, double fs, double omega);
 
-int read_from_file(const char* name, float* buffer, int start);
-void write_to_file(const char* name, float* data, int size);
+int read_from_file(const char* name, double* buffer, int start);
+void write_to_file(const char* name, double* data, int size);
 
 
 #endif
